# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Synapse Hospitality API",
    "description": "Especificação Oficial OpenAPI 3.0 para a Plataforma Synapse Hospitality (PMS, CRM, AI Orchestrator, Governança, Manutenção, Recepção e Integrações).",
    "termsOfService": "https://synapsehospitality.com.br/terms",
    "contact": {
      "name": "Suporte Técnico Synapse Hospitality",
      "email": "suporte@synapsehospitality.com.br",
      "url": "https://synapsehospitality.com.br"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://synapsehospitality.com.br/license"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1 - Versão Oficial Padronizada"
    },
    {
      "url": "/api",
      "description": "API Base - Endpoints da Aplicação"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "OrganizationHeader": [],
      "PropertyHeader": []
    }
  ],
  "paths": {
...
    "/strategy/dashboard": {
      "get": {
        "tags": ["Strategic Simulation & Explainable AI"],
        "summary": "Retorna o dashboard consolidado de simulações estratégicas, potencial financeiro e métricas em memória",
        "operationId": "getStrategyDashboard",
        "parameters": [
          { "$ref": "#/components/parameters/XOrganizationId" },
          { "$ref": "#/components/parameters/XPropertyId" }
        ],
        "responses": {
          "200": { "description": "Dashboard de estratégia retornado com sucesso" }
        }
      }
    },
    "/strategy/scenarios": {
      "get": {
        "tags": ["Strategic Simulation & Explainable AI"],
        "summary": "Retorna a lista de cenários de simulação What If disponíveis com Explainable AI completa",
        "operationId": "getStrategyScenarios",
        "parameters": [
          { "$ref": "#/components/parameters/XOrganizationId" },
          { "$ref": "#/components/parameters/XPropertyId" }
        ],
        "responses": {
          "200": { "description": "Lista de cenários de simulação retornada com sucesso" }
        }
      }
    },
    "/strategy/simulate": {
      "post": {
        "tags": ["Strategic Simulation & Explainable AI"],
        "summary": "Executa uma simulação tática/estratégica customizada em memória sem alterar nenhum dado no banco",
        "operationId": "simulateCustomStrategy",
        "parameters": [
          { "$ref": "#/components/parameters/XOrganizationId" },
          { "$ref": "#/components/parameters/XPropertyId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scenarioType": { "type": "string", "example": "adr_increase" },
                  "adrIncreasePercent": { "type": "number", "example": 10 },
                  "cancellationReductionPercent": { "type": "number", "example": 30 },
                  "conversionIncreasePercent": { "type": "number", "example": 15 },
                  "occupancyIncreasePercent": { "type": "number", "example": 8 },
                  "customName": { "type": "string", "example": "Simulação Alta Temporada 2026" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Simulação executada com sucesso em memória" }
        }
      }
    },
    "/strategy/summary": {
      "get": {
        "tags": ["Strategic Simulation & Explainable AI"],
        "summary": "Retorna o resumo agregado das simulações estratégicas para a IA (ContextService)",
        "operationId": "getStrategySummary",
        "parameters": [
          { "$ref": "#/components/parameters/XOrganizationId" },
          { "$ref": "#/components/parameters/XPropertyId" }
        ],
        "responses": {
          "200": { "description": "Resumo de estratégia retornado com sucesso" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Autenticação via Token JWT Bearer"
      },
      "OrganizationHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-organization-id",
        "description": "Identificador do Tenant (Organização)"
      },
      "PropertyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-property-id",
        "description": "Identificador do Tenant (Propriedade Hoteleira)"
      }
    },
...
  }
}